Extract doc fragment from openssl_csr.
